# docs/swagger.yaml
basePath: /
definitions:
  entities.Book:
    properties:
      author:
        type: string
      available_date:
        type: string
      category:
        $ref: '#/definitions/entities.Category'
      category_id:
        type: integer
      description:
        type: string
      discount:
        $ref: '#/definitions/entities.Discount'
      discount_id:
        type: integer
      id:
        type: integer
      isbn:
        type: string
      language:
        type: string
      link:
        type: string
      preorder:
        type: boolean
      price:
        type: number
      publisher:
        $ref: '#/definitions/entities.Publisher'
      publisher_id:
        type: integer
      rating:
        type: number
      title:
        type: string
    type: object
  entities.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  entities.Discount:
    properties:
      description:
        type: string
      id:
        type: integer
      percentage:
        type: number
      title:
        type: string
    type: object
...
  entities.Review:
    properties:
      book:
        $ref: '#/definitions/entities.Book'
      book_id:
        type: integer
      comment:
        type: string
      id:
        type: integer
      rating:
        type: string
      user:
        $ref: '#/definitions/entities.User'
      user_id:
        type: integer
    type: object
  entities.User:
    properties:
      address:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      password:
        type: string
      phone:
        type: string
      reg_date:
        type: string
      role:
        type: string
    type: object
...
host: project11a-backend-the-seal-division.onrender.com
info:
  contact:
    email: '[email protected]'
    name: API Support
    url: http://www.swagger.io/support
  description: REST API для онлайн-магазина книг
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: BookStore Backend API
  version: 1.0.0
paths:
  /books:
    post:
      consumes:
      - application/json
      description: Создание новой книги в базе данных
      parameters:
      - description: Данные книги
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/entities.Book'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Book'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавить книгу
      tags:
      - books
  /books/{id}:
    delete:
      consumes:
      - application/json
      description: Удаление книги по ID
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить книгу
      tags:
      - books
    get:
      consumes:
      - application/json
      description: Получение информации о книге по ID
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Book'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить книгу по ID
      tags:
      - books
    put:
      consumes:
      - application/json
      description: Обновление информации о книге
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные книги
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/entities.Book'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Book'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить книгу
      tags:
      - books
...
  /books/authors/{author}:
    get:
      consumes:
      - application/json
      description: Получение списка книг определенного автора
      parameters:
      - description: Автор
        in: path
        name: author
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Book'
            type: array
      summary: Получить книги по автору
      tags:
      - books
  /books/categories/{category}:
    get:
      consumes:
      - application/json
      description: Получение списка книг определенной категории
      parameters:
      - description: Категория книг
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Book'
            type: array
      summary: Получить книги по категории
      tags:
      - books
  /books/publishers/{publisher}:
    get:
      consumes:
      - application/json
      description: Получение списка книг определенного издателя
      parameters:
      - description: Издатель
        in: path
        name: publisher
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Book'
            type: array
      summary: Получить книги по издателю
      tags:
      - books
  /books/reviews/{id}:
    delete:
      consumes:
      - application/json
      description: Удаление отзыва по ID
      parameters:
      - description: ID отзыва
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить отзыв
      tags:
      - reviews
    get:
      consumes:
      - application/json
      description: Получение информации об отзыве по ID
      parameters:
      - description: ID отзыва
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Review'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить отзыв
      tags:
      - reviews
    post:
      consumes:
      - application/json
      description: Создание нового отзыва
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: integer
      - description: Данные отзыва
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/entities.Review'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Review'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавить отзыв
      tags:
      - reviews
    put:
      consumes:
      - application/json
      description: Обновление информации об отзыве
      parameters:
      - description: ID отзыва
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные отзыва
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/entities.Review'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Review'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить отзыв
      tags:
      - reviews
  /books/reviews/book/{id}:
    get:
      consumes:
      - application/json
      description: Получение списка отзывов по ID книги
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Review'
            type: array
      summary: Получить отзывы книги
      tags:
      - reviews
  /categories:
    post:
      consumes:
      - application/json
      description: Создание новой категории
      parameters:
      - description: Данные категории
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/entities.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавить категорию
      tags:
      - categories
  /categories/{id}:
    get:
      consumes:
      - application/json
      description: Получение информации о категории по ID
      parameters:
      - description: ID категории
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Category'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить категорию
      tags:
      - categories
...
  /discounts:
    post:
      consumes:
      - application/json
      description: Создание новой скидки
      parameters:
      - description: Данные скидки
        in: body
        name: discount
        required: true
        schema:
          $ref: '#/definitions/entities.Discount'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Discount'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавить скидку
      tags:
      - discounts
  /discounts/{id}:
    patch:
      consumes:
      - application/json
      description: Обновление информации о скидке
      parameters:
      - description: ID скидки
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные скидки
        in: body
        name: discount
        required: true
        schema:
          $ref: '#/definitions/entities.Discount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Discount'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить скидку
      tags:
      - discounts
...
  /login:
    post:
      consumes:
      - application/json
      description: Аутентификация пользователя по email и паролю
      parameters:
      - description: Данные для входа
        in: body
        name: credentials
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход в систему
      tags:
      - auth
  /orders:
    post:
      consumes:
      - application/json
      description: Создание нового заказа
      parameters:
      - description: Данные заказа
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/entities.Orders'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Orders'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать заказ
      tags:
      - orders
...
  /orders/user/{id}:
    get:
      consumes:
      - application/json
      description: Получение списка заказов по ID пользователя
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Orders'
            type: array
      summary: Получить заказы пользователя
      tags:
      - orders
  /publishers:
    post:
      consumes:
      - application/json
      description: Создание нового издателя
      parameters:
      - description: Данные издателя
        in: body
        name: publisher
        required: true
        schema:
          $ref: '#/definitions/entities.Publisher'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Publisher'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавить издателя
      tags:
      - publishers
  /publishers/{id}:
    get:
      consumes:
      - application/json
      description: Получение информации об издателе по ID
      parameters:
      - description: ID издателя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Publisher'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить издателя
      tags:
      - publishers
...
schemes:
- https
- http
swagger: "2.0"